new splat gen flow working
